feat: Implementar paginação e filtros na busca de categorias

+++ src/services/serviceCategoria.js
@@ -2,17 +2,41 @@ const tabelaCategoria = require('../models/tabelaCategoria');
 const resposta = require('../responses');
 
 
+
 const getCategorias = async (req, res) => {
     try {
-        const categorias = await tabelaCategoria.findAll();
-        if(!categorias) return resposta.badRequest(res,'campo vazio')
+        const { limit = 12, page = 1, fields, use_in_menu } = req.query;
+        const queryOptions = {};
+        if (use_in_menu) {
+            queryOptions.where = { use_in_menu: use_in_menu === 'true' };
+        }
+        if (fields) {
+            queryOptions.attributes = fields.split(',');
+        }
+        if (parseInt(limit) !== -1) {
+            queryOptions.limit = parseInt(limit);
+            queryOptions.offset = (parseInt(page) - 1) * parseInt(limit);
+        }
 
-        resposta.success(res,'Categorias encontradas',categorias)
-            
+        const categorias = await tabelaCategoria.findAndCountAll(queryOptions);
+        const response = {
+            data: categorias.rows,
+            total: categorias.count,
+            limit: parseInt(limit),
+            page: parseInt(limit) === -1 ? null : parseInt(page)
+        };
+
+        if (categorias.rows.length > 0) {
+            return res.status(200).json(response);
+        } else {
+            return res.status(400).json({ mensagem: 'Nenhuma categoria encontrada' });
+        }
     } catch (error) {
-        resposta.InternalServerError(res,'Ocorreu um erro ao procura as categorias')
+        return res.status(400).json({ mensagem: 'Erro na requisição', erro: error.message });
     }
-}
+};
+
+
 
 const getCategoriaId = async (req, res) => {
     const id = req.params.id;
@@ -26,32 +50,42 @@ const getCategoriaId = async (req, res) => {
         resposta.InternalServerError(res,'Ocorreu um erro ao procura a categoria')
     }
 }
-
-//post
-
 const postCategoria = async (req, res) => {
     const { name,slug,use_in_menu } = req.body;
-    if(!name || !slug|| !use_in_menu) return resposta.badRequest(res,'Categorias não encotrada')
+    const token = req.headers.authorization;
+
+    if (!token) {
+        return resposta.unauthorized(res, 'Token de autorização não fornecido');
+    }
+
+    if (!name || !slug || !use_in_menu) {
+        resposta.badRequest(res, 'Todos os campos são obrigatórios');
+    }
     try {
         const novaCategoria = await tabelaCategoria.create({
             name: name,
             slug: slug,
             use_in_menu: use_in_menu,
         });
-
-        resposta.success(res,'Categoria cadastrada com sucesso',categoria)
-            
+        if(novaCategoria){
+            resposta.created(res,novaCategoria)
+        }else{
+            resposta.badRequest(res,'Erro ao criar categoria')
+        }
     } catch (error) {
-        resposta.InternalServerError(res,'Ocorreu um erro na criação de uma nova categoria')
+        res.json(error) 
     }
 }
 
-//put
 
 const putCategoria = async (req, res) => {
     const id = req.params.id;
+    const token = req.headers.authorization;
     const { name, slug, use_in_menu } = req.body;
-        //FALTA O TOKEN PARA O 401
+        //TOKEN PARA O 401
+        if (!token) {
+            return resposta.unauthorized(res, 'Token de autorização não fornecido');
+        }
 
     try {
         const AttCategoria = await tabelaCategoria.update({
@@ -75,7 +109,12 @@ const putCategoria = async (req, res) => {
 
 const deleteCategoria = async (req, res) => {
     const id = req.params.id;
+    const token = req.headers.authorization;
     //FALTA O TOKEN PARA O 401
+    if (!token) {
+        return resposta.unauthorized(res, 'Token de autorização não fornecido');
+    }
+
     try {
             const categoriaDelet = await tabelaCategoria.destroy({
                 where: {
